requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
transformers
torch
faiss-cpu
requests
pytest
semantic-kernel
pydantic
python-dotenv
langchain-text-splitters